c++ refactored

+++ src/copyright/agent/copyrightMatch.cc
@@ -9,24 +9,29 @@
  * You should have received a copy of the GNU General Public License along with this program; if not, write to the Free Software Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  */
 
+#include <stddef.h>
 #include "copyrightMatch.hpp"
 
-CopyrightMatch::CopyrightMatch(std::string content, std::string type, unsigned int start, unsigned int length):
- content(content), start(start), length(length), type(type)
-{}
+CopyrightMatch::CopyrightMatch(std::string content, std::string type, unsigned int start, unsigned int length) :
+        content(content), start(start), length(length), type(type)
+{
+}
 
-CopyrightMatch::CopyrightMatch(std::string content, std::string type, unsigned int start):
- content(content), start(start), length(content.length()), type(type)
-{}
+CopyrightMatch::CopyrightMatch(std::string content, std::string type, unsigned int start) :
+        content(content), start(start), length(content.length()), type(type)
+{
+}
 
-CopyrightMatch::~CopyrightMatch(){};
+CopyrightMatch::~CopyrightMatch()
+{
+};
 
-unsigned CopyrightMatch::getStart() const
+size_t CopyrightMatch::getStart() const
 {
   return start;
 }
 
-unsigned CopyrightMatch::getLength() const
+size_t CopyrightMatch::getLength() const
 {
   return length;
 }
@@ -43,29 +48,29 @@ const std::string CopyrightMatch::getType() const
   return type;
 }
 
-bool operator==(const CopyrightMatch& first, const CopyrightMatch& other)
+bool operator ==(const CopyrightMatch& first, const CopyrightMatch& other)
 {
   return (first.getContent() == other.getContent()) &&
-         (first.getType() == other.getType()) &&
-         (first.getStart() == other.getStart()) &&
-         (first.getLength() == other.getLength());
+          (first.getType() == other.getType()) &&
+          (first.getStart() == other.getStart()) &&
+          (first.getLength() == other.getLength());
 }
 
-std::ostream& operator<<(std::ostream& os, const CopyrightMatch& match)
+std::ostream& operator <<(std::ostream& os, const CopyrightMatch& match)
 {
-  unsigned start = match.getStart();
-  unsigned end = start + match.getLength();
+  size_t start = match.getStart();
+  size_t end = start + match.getLength();
 
   os << "["
           << start << ":" << end << ":" << match.getType()
-     << "] '" <<
-              match.getContent()
-       << "'";
+          << "] '" <<
+          match.getContent()
+                  << "'";
 
   return os;
 }
 
-std::ostream& operator<<(std::ostream& os, const std::vector< CopyrightMatch >& matches)
+std::ostream& operator <<(std::ostream& os, const std::vector<CopyrightMatch>& matches)
 {
   typedef std::vector<CopyrightMatch>::const_iterator cpm;
   for (cpm it = matches.begin(); it != matches.end(); ++it)
