+++ src/copyright/agent/copyrightMatch.hpp
@@ -19,24 +19,24 @@
 class CopyrightMatch
 {
 public:
-    CopyrightMatch(std::string content, std::string type, unsigned start, unsigned length);
-    CopyrightMatch(std::string content, std::string type, unsigned start);
-    ~CopyrightMatch();
-    const std::string getType() const;
-    const std::string getContent() const;
-    unsigned getStart() const;
-    unsigned getLength() const;
+  CopyrightMatch(std::string content, std::string type, unsigned start, unsigned length);
+  CopyrightMatch(std::string content, std::string type, unsigned start);
+  ~CopyrightMatch();
+  const std::string getType() const;
+  const std::string getContent() const;
+  size_t getStart() const;
+  size_t getLength() const;
 private:
-    std::string content;
-    unsigned start;
-    unsigned length;
-    std::string type;
+  std::string content;
+  unsigned start;
+  size_t length;
+  std::string type;
 };
 
-std::ostream& operator<<(std::ostream& os, const CopyrightMatch& match);
-bool operator==(const CopyrightMatch& first, const CopyrightMatch& other);
+std::ostream& operator <<(std::ostream& os, const CopyrightMatch& match);
+bool operator ==(const CopyrightMatch& first, const CopyrightMatch& other);
 
-std::ostream& operator<<(std::ostream& os, const std::vector<CopyrightMatch>& matches);
+std::ostream& operator <<(std::ostream& os, const std::vector<CopyrightMatch>& matches);
 
 
 #endif // COPYRIGHTMATCH_H
